+++ src/lib/getIdentifierName.js
@@ -5,7 +5,7 @@ module.exports = function (element, param, options) {
     if (param) { return param; }
     if (options && options.assigned) { 
         count += 1;
-        return "$SIDE_EFFECT_" + count; 
+        return "$ASSIGNED_" + count; 
     }
     return string.camelize(element.replace(".", "-"));
 }