+++ test/fixture/cli/named-assigned-side-effects/output.js
@@ -1,5 +1,5 @@
 import Marionette from 'backbone.marionette';
-import $SIDE_EFFECT_1 from 'side-effect';
+import $ASSIGNED_1 from 'side-effect';
 export default Marionette.Object.extend({
     initialize: function (options) {
         this.window.hello();