+++ test/spec/index.js
@@ -213,7 +213,7 @@ test("it should convert require sugar with side effects correctly", t => {
     t.truthy(convert("define/require-sugar-with-side-effect"));
 });
 
-test.skip("it should convert require with property assignment correctly", t => {
+test("it should convert require with property assignment correctly", t => {
     t.truthy(convert("define/require-sugar-with-property-assignment"));
 });